[Chore] 챌린지 조회, 상세 조회 시 사용자 참여수로 반환하도록 수정

+++ src/main/java/dev/book/challenge/entity/Challenge.java
@@ -98,7 +98,7 @@ public class Challenge extends BaseTimeEntity {
         });
     }
 
-    public void isOver(long countParticipants) {
+    public void isOver(int countParticipants) {
         if (countParticipants >= this.capacity) {
             throw new ChallengeException(CHALLENGE_CAPACITY_FULL);
         }
